reorganize tests for mars-account-nft
