+++ contracts/account-nft/src/lib.rs
@@ -4,3 +4,5 @@ pub mod execute;
 pub mod migrations;
 pub mod query;
 pub mod state;
+#[cfg(test)]
+mod tests;
