+++ contracts/account-nft/src/tests/mint.rs
@@ -2,15 +2,13 @@ use cosmwasm_std::Addr;
 use cw721::OwnerOfResponse;
 use cw721_base::{ContractError::Ownership, OwnershipError::NotOwner};
 use cw_multi_test::Executor;
-use mars_account_nft::error::{ContractError, ContractError::BaseError};
 use mars_types::{
     account_nft::{ExecuteMsg, QueryMsg::OwnerOf},
     health::AccountKind,
 };
 
-use crate::helpers::{below_max_for_burn, MockEnv};
-
-pub mod helpers;
+use super::helpers::{below_max_for_burn, MockEnv};
+use crate::error::{ContractError, ContractError::BaseError};
 
 #[test]
 fn id_incrementer() {
