+++ contracts/account-nft/src/tests/mod.rs
@@ -1,0 +1,8 @@
+mod burn_allowance;
+mod burn_empty_accounts;
+mod helpers;
+mod instantiate;
+mod migration;
+mod mint;
+mod proposed_minter;
+mod update_config;
