--- contracts/account-nft/tests/helpers/mod.rs
@@ -1,6 +1,0 @@
-pub use self::{health_responses::*, mock_contracts::*, mock_env::*, mock_env_builder::*};
-
-mod health_responses;
-mod mock_contracts;
-mod mock_env;
-mod mock_env_builder;
